// apps/frontend/src/app/enums/news-types.enum.ts
export enum NewsTypes {
  OFFLINE_START = 'offline_start',
  OFFLINE_RESULTS = 'offline_results',
  ONLINE_ANNOUNCE = 'online_announce',
  ONLINE_RESULTS = 'online_results',
  MULTICUP_RESULTS = 'multicup_results',
  SIMPLE = 'text',
  DFWC_RESULTS = 'dfwc_round_results',
  REFLEX_OFFLINE_START = 'reflex_offline_start',
  REFLEX_OFFLINE_RESULTS = 'reflex_offline_results',
}